+++ src/__tests__/validateInvocation.test.ts
@@ -1,9 +1,11 @@
-import { createMockExecutionContext } from '@jupiterone/integration-sdk/testing';
+import { IntegrationConfig } from '../types';
+
+import { createMockExecutionContext } from '@jupiterone/integration-sdk-testing';
 
 import validateInvocation from '../validateInvocation';
 
 test('rejects if apiKey is not present', async () => {
-  const context = createMockExecutionContext();
+  const context = createMockExecutionContext<IntegrationConfig>();
   context.instance.config['apiKey'] = undefined;
 
   await expect(validateInvocation(context)).rejects.toThrow(
@@ -12,7 +14,7 @@ test('rejects if apiKey is not present', async () => {
 });
 
 test('rejects if unable to hit provider apis', async () => {
-  const context = createMockExecutionContext();
+  const context = createMockExecutionContext<IntegrationConfig>();
   context.instance.config = { apiKey: 'test' };
 
   await expect(validateInvocation(context)).rejects.toThrow(