+++ src/collector/ServicesClient.ts
@@ -1,10 +1,10 @@
 import meraki = require('meraki');
-import { promisfy } from 'promisfy';
 import {
   MerakiOrganization,
   MerakiNetwork,
   MerakiVlan,
   MerakiAdminUser,
+  MerakiClient,
   MerakiDevice,
   MerakiSamlRole,
   MerakiSSID,
@@ -28,74 +28,79 @@ export class ServicesClient {
    * Get Organizations
    */
   async getOrganizations(): Promise<MerakiOrganization[]> {
-    const getOrganizations = promisfy(
-      meraki.OrganizationsController.getOrganizations,
-    );
-    const res: object[] = await getOrganizations();
-    return res[0] as MerakiOrganization[];
+    const res: object[] = await meraki.OrganizationsController.getOrganizations();
+    return res as MerakiOrganization[];
   }
 
   /**
    * Get Admin Users of an Organization
    */
   async getAdmins(organizationId: string): Promise<MerakiAdminUser[]> {
-    const getOrganizationAdmins = promisfy(
-      meraki.AdminsController.getOrganizationAdmins,
+    const res: object[] = await meraki.AdminsController.getOrganizationAdmins(
+      organizationId,
     );
-    const res: object[] = await getOrganizationAdmins(organizationId);
-    return res[0] as MerakiAdminUser[];
+    return res as MerakiAdminUser[];
   }
 
   /**
    * Get SAML Roles of an Organization
    */
   async getSamlRoles(organizationId: string): Promise<MerakiSamlRole[]> {
-    const getOrganizationSamlRoles = promisfy(
-      meraki.SAMLRolesController.getOrganizationSamlRoles,
+    const res: object[] = await meraki.SAMLRolesController.getOrganizationSamlRoles(
+      organizationId,
     );
-    const res: object[] = await getOrganizationSamlRoles(organizationId);
-    return res[0] as MerakiSamlRole[];
+    return res as MerakiSamlRole[];
   }
 
   /**
    * Get Networks of an Organization
    */
   async getNetworks(organizationId: string): Promise<MerakiNetwork[]> {
-    const getOrganizationNetworks = promisfy(
-      meraki.NetworksController.getOrganizationNetworks,
+    const res: object[] = await meraki.NetworksController.getOrganizationNetworks(
+      {
+        organizationId,
+      },
     );
-    const res: object[] = await getOrganizationNetworks({
-      organizationId,
-    });
-    return res[0] as MerakiNetwork[];
+    return res as MerakiNetwork[];
   }
 
   /**
    * Get VLANs in a Network
    */
   async getVlans(networkId: string): Promise<MerakiVlan[]> {
-    const getNetworkVlans = promisfy(meraki.VlansController.getNetwork_vlans);
-    const res: object[] = await getNetworkVlans(networkId);
-    return res[0] as MerakiVlan[];
+    const res: object[] = await meraki.VlansController.getNetwork_vlans(
+      networkId,
+    );
+    return res as MerakiVlan[];
   }
 
   /**
    * Get Devices in a Network
    */
   async getDevices(networkId: string): Promise<MerakiDevice[]> {
-    const getNetworkDevices = promisfy(
-      meraki.DevicesController.getNetworkDevices,
+    const res: object[] = await meraki.DevicesController.getNetworkDevices(
+      networkId,
     );
-    const res: object[] = await getNetworkDevices(networkId);
-    return res[0] as MerakiDevice[];
+    return res as MerakiDevice[];
+  }
+
+  /**
+   * Get Clients in a Network
+   */
+  async getClients(networkId: string): Promise<MerakiClient[]> {
+    const res: object[] = await meraki.ClientsController.getNetworkClients({
+      networkId,
+    });
+    return res as MerakiClient[];
   }
 
   /**
    * Get SSIDs of a Wireless Network
    */
   async getSSIDs(networkId: string): Promise<MerakiSSID[]> {
-    const getNetworkSSIDs = promisfy(meraki.SsidsController.getNetwork_ssids);
-    const res: object[] = await getNetworkSSIDs(networkId);
-    return res[0] as MerakiSSID[];
+    const res: object[] = await meraki.SsidsController.getNetwork_ssids(
+      networkId,
+    );
+    return res as MerakiSSID[];
   }
 }