+++ src/index.ts
@@ -1,12 +1,13 @@
-import { IntegrationInvocationConfig } from '@jupiterone/integration-sdk';
+import { IntegrationInvocationConfig } from '@jupiterone/integration-sdk-core';
 
 import instanceConfigFields from './instanceConfigFields';
+import fetchClients from './steps/fetch-clients';
+import fetchResources from './steps/fetch-resources';
+import { IntegrationConfig } from './types';
 import validateInvocation from './validateInvocation';
 
-import synchronize from './steps/synchronize';
-
-export const invocationConfig: IntegrationInvocationConfig = {
+export const invocationConfig: IntegrationInvocationConfig<IntegrationConfig> = {
   instanceConfigFields,
   validateInvocation,
-  integrationSteps: [synchronize],
+  integrationSteps: [fetchResources, fetchClients],
 };