+++ CHANGELOG.md
@@ -8,4 +8,21 @@ and this project adheres to
 
 ## [Unreleased]
 
+## 1.3.2 - 2020-06-01
+
 ### Changed
+
+- Upgraded to `@jupiterone/integration-sdk@^3.1.0` to get TypeScript type fixes.
+
+## 1.3.1 - 2020-06-01
+
+### Fixed
+
+- Define `IntegrationConfig` type and add as generic parameter as required.
+
+## 1.3.0 - 2020-05-31
+
+### Added
+
+- Use mapped relationships to connect network entities to clients discovered on
+  the network.